refactor: group host and desktop modules by concern
core/host gains remote-discovery.ts, holding the two questions asked when reaching
a console over the network: which local interfaces can host one, and whether the
TLS identity at an address matches the link.

fleet-desktop gains boot-dialogs.ts (a failed sidecar boot and an unadoptable
console lock both end in a dialog and an exit) and console-links.ts (which origins
count as the console, and how a fleet:// link is spotted in argv).

stale.ts and health.ts are deliberately left apart. Merging them built and typed
fine but broke the dts rollup: health.ts carries no fs import, so folding it in
with stale.ts made `typeof fs` reachable from the cli declaration entry, and
node:fs has no default export to bundle there. Keeping them split costs one file
and avoids reshaping a published type to suit a file layout.

packages/CLAUDE.md listed five of the seven packages; core-ai-gateway and
fleet-analyst were missing from the ownership route.
