+++ runtime/fleet-console/tests/console-settings.test.ts
@@ -4,7 +4,7 @@ import os from "node:os";
 
 import { afterEach, describe, expect, it } from "vitest";
 
-import { listRemoteInterfaces } from "../core/host/remote-interfaces.js";
+import { listRemoteInterfaces } from "../core/host/remote-discovery.js";
 import {
   createConsoleSettingsStore,
   sanitizeConsoleSettingsData,
