+++ runtime/fleet-desktop/src/boot-dialogs.ts
@@ -1,3 +1,11 @@
+// Fatal startup conditions the user has to see: the sidecar would not come up, or a
+// console this Desktop cannot adopt already owns the lock. Both end in a dialog and
+// then an exit, so they share this file rather than each owning one.
+
+
+
+// ─── boot failure ──────────────────────────────────────────────────────────────
+
 /**
  * 부팅이 끝내 실패했을 때 사용자에게 남기는 말.
  *
@@ -43,5 +51,47 @@ export function showBootFailureAndExit(error: unknown, dependencies: BootFailure
     // 다이얼로그를 띄우지 못하는 상황이라도 종료 처리는 그대로 진행한다.
   } finally {
     dependencies.exit(1);
+  }
+}
+
+// ─── console already owned ─────────────────────────────────────────────────────
+
+export interface ConsoleConflictDialogOptions {
+  readonly buttons: string[];
+  readonly cancelId: number;
+  readonly defaultId: number;
+  readonly detail: string;
+  readonly message: string;
+  readonly title: string;
+  readonly type: "warning";
+}
+
+export interface ConsoleConflictHandlerDependencies {
+  readonly quit: () => void;
+  readonly showMessageBox: (options: ConsoleConflictDialogOptions) => Promise<unknown>;
+}
+
+const CONSOLE_CONFLICT_ERRORS = new Set(["cli_daemon_requires_confirmation", "console_lock_foreign_process_unhealthy"]);
+const CONSOLE_CONFLICT_DIALOG: ConsoleConflictDialogOptions = {
+  type: "warning",
+  title: "Fleet Console is already running",
+  message: "Fleet Console is already running.",
+  detail: "Stop or quit the running Fleet Console before opening Fleet Console Desktop again.",
+  buttons: ["OK"],
+  defaultId: 0,
+  cancelId: 0,
+};
+
+export function isConsoleConflict(error: unknown): boolean {
+  return error instanceof Error && CONSOLE_CONFLICT_ERRORS.has(error.message);
+}
+
+export async function showConsoleConflictAndQuit(dependencies: ConsoleConflictHandlerDependencies): Promise<void> {
+  try {
+    await dependencies.showMessageBox(CONSOLE_CONFLICT_DIALOG);
+  } catch {
+    // The Desktop must still quit if Electron cannot display the acknowledgement dialog.
+  } finally {
+    dependencies.quit();
   }
 }
