+++ runtime/fleet-desktop/src/main.ts
@@ -6,8 +6,7 @@ import { fileURLToPath } from "node:url";
 import { app, BrowserWindow, dialog, Menu, Notification, screen, session, shell, Tray, WebContentsView } from "electron";
 
 import { createDesktopLifecycle } from "./app-lifecycle.js";
-import { showBootFailureAndExit } from "./boot-failure.js";
-import { isConsoleConflict, showConsoleConflictAndQuit } from "./console-conflict.js";
+import { isConsoleConflict, showBootFailureAndExit, showConsoleConflictAndQuit } from "./boot-dialogs.js";
 import { createConsoleControls } from "./console-controls.js";
 import { handOffWindowToConsole } from "./console-handoff.js";
 import { createHydratedDesktopEnvironment, resolveDesktopUserDataDirectory } from "./environment.js";
@@ -16,10 +15,9 @@ import { applyDesktopDockIcon, applyDesktopIdentity } from "./identity.js";
 import { createLaunchController, type RuntimeEntryState } from "./launch-controller.js";
 import { createDesktopNotifier } from "./desktop-notices.js";
 import { createHostPickerView } from "./host-picker-view.js";
-import { isRemoteConsoleOrigin } from "./console-origin.js";
+import { findAccessLinkArgument, FLEET_PROTOCOL, isFleetProtocolLink, isRemoteConsoleOrigin } from "./console-links.js";
 import { installRemoteCertificatePins } from "./remote-access.js";
 import { consoleTarget, createRemoteBridge, type RemoteBridge } from "./remote-bridge.js";
-import { findAccessLinkArgument, FLEET_PROTOCOL, isFleetProtocolLink } from "./fleet-protocol.js";
 import { createDesktopLogger, describeError, type DesktopLogger } from "./logging.js";
 import { createDesktopThemeSynchronizer } from "./desktop-theme-sync.js";
 import { createDesktopUpdateSynchronizer } from "./desktop-update-sync.js";
